move refresh logic to separate svc
