+++ src/app/action-list/action-list.component.ts
@@ -3,7 +3,7 @@ import { ActionItem } from '../../domain/action-item';
 import { GithubService } from '../../github/services/github.service';
 import { VstsService } from '../../github/services/vsts.service';
 import { JenkinsService } from '../../jenkins/services/jenkins.service';
-import { ACTION_ITEM_POLLING_INTERVAL_IN_MS } from '../../config/app-config-constants';
+import { ACTION_ITEM_POLLING_INTERVAL_IN_MS } from '../app.constants';
 import { ConfigService } from '../../app/config.service';
 import { NotificationsService } from '../../notifications/services/notifications.service';
 
@@ -35,7 +35,6 @@ export class ActionListComponent implements OnInit {
       this.githubService.loadRepos().then(() => {
         this.getActionItemsList();
         setInterval(() => {
-          this.configService.checkForRefresh();
           this.getActionItemsList();
         }, ACTION_ITEM_POLLING_INTERVAL_IN_MS);
       });
