+++ src/app/app.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { APP_LABELS } from './app.constants';
+import { RefreshService } from './refresh.service';
 
 @Component({
   selector: 'mf-app-root',
@@ -8,4 +9,7 @@ import { APP_LABELS } from './app.constants';
 })
 export class AppComponent {
   title = APP_LABELS.TITLE;
+
+  constructor(private refreshService: RefreshService) {
+  }
 }
